Change ScalePrecisionValidator to to check digits to left of decimal point
Change the behavior of ScalePrecisionValidator to behave like SQL Server / MySQL / Oracle / etc. where the precision is the total number of digits and the scale is the number of digits to the right of the decimal. The difference between the precision and the scale should be the number of digits allowed on the left side of the decimal point.

#1008