Update including-rules.md
If the included validator is alphabetically/hierarchically placed below the root validator then only included validator is getting called and validated on the model while root is completely ignored. This issue was reported on the stackoverflow.com but nobody had given solution to this problem.
I thought this small text edit might help developers to solve their problem with include() validator.